# FitRest/FitRest/OpenAPIs/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "FitRestAPI",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Days": {
      "get": {
        "tags": [
          "Days"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Day"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Day"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Day"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Days"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Day"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Day"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Day"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Day"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Day"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Day"
                }
              }
            }
          }
        }
      }
    },
    "/api/Days/{id}": {
      "get": {
        "tags": [
          "Days"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Day"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Day"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Day"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Days"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Day"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Day"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Day"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Days"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Meal": {
      "get": {
        "tags": [
          "Meal"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Meal"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Meal"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Meal"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Meal"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Meal"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Meal"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Meal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Meal"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meal"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meal"
                }
              }
            }
          }
        }
      }
    },
    "/api/Meal/{id}": {
      "get": {
        "tags": [
          "Meal"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Meal"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meal"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Meal"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Meal"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Meal"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Meal"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Meal"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Meal"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/MealIngredient": {
      "get": {
        "tags": [
          "MealIngredient"
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealIngredient"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealIngredient"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/MealIngredient"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "MealIngredient"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealIngredient"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MealIngredient"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MealIngredient"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MealIngredient"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealIngredient"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealIngredient"
                }
              }
            }
          }
        }
      }
    },
    "/api/MealIngredient/{id}": {
      "get": {
        "tags": [
          "MealIngredient"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/MealIngredient"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealIngredient"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealIngredient"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "MealIngredient"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MealIngredient"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/MealIngredient"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/MealIngredient"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "MealIngredient"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Activity": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "caloriesPerMinute": {
            "type": "integer",
            "format": "int32"
          },
          "activityDuration": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityWithDuration"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ActivityWithDuration": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "durationInMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "intensity": {
            "type": "number",
            "format": "double"
          },
          "idDay": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "$ref": "#/components/schemas/Day"
          },
          "idActivity": {
            "type": "integer",
            "format": "int32"
          },
          "activity": {
            "$ref": "#/components/schemas/Activity"
          }
        },
        "additionalProperties": false
      },
      "Day": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "caloriesGoal": {
            "type": "number",
            "format": "double"
          },
          "carbohydratesGoal": {
            "type": "number",
            "format": "double"
          },
          "proteinGoal": {
            "type": "number",
            "format": "double"
          },
          "fatsGoal": {
            "type": "number",
            "format": "double"
          },
          "activityWithDuration": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActivityWithDuration"
            },
            "nullable": true
          },
          "sleep": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Sleep"
            },
            "nullable": true
          },
          "meals": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Meal"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Meal": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "timeOfEating": {
            "type": "string",
            "format": "date-time"
          },
          "mealIngredientWithPortions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealIngredientWithPortion"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MealIngredient": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "carbohydrates": {
            "type": "integer",
            "format": "int32"
          },
          "proteins": {
            "type": "integer",
            "format": "int32"
          },
          "fats": {
            "type": "integer",
            "format": "int32"
          },
          "calories": {
            "type": "integer",
            "format": "int32"
          },
          "portionName": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MealIngredientWithPortion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "portionSize": {
            "type": "number",
            "format": "double"
          },
          "idMeal": {
            "type": "integer",
            "format": "int32"
          },
          "meal": {
            "$ref": "#/components/schemas/Meal"
          },
          "idMealIngredient": {
            "type": "integer",
            "format": "int32"
          },
          "mealIngredient": {
            "$ref": "#/components/schemas/MealIngredient"
          }
        },
        "additionalProperties": false
      },
      "Sleep": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "modifiedDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "startTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "endTime": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "durationInMinutes": {
            "type": "integer",
            "format": "int32"
          },
          "durationInMinutesOfDeepSleep": {
            "type": "integer",
            "format": "int32"
          },
          "ratingOfSleep": {
            "type": "integer",
            "format": "int32"
          },
          "idDay": {
            "type": "integer",
            "format": "int32"
          },
          "day": {
            "$ref": "#/components/schemas/Day"
          }
        },
        "additionalProperties": false
      }
    }
  }
}